Passage des float en double précision dans le stockage intermédiaire

+++ database/migrations/2023_10_02_220237_create_variant_attributes_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('value_str')->nullable();
             $table->text('value_txt')->nullable();
             $table->integer('value_int')->nullable();
-            $table->float('value_float')->nullable();
+            $table->double('value_float')->nullable();
             $table->timestamps();
         });
     }
